changed Pipeline.py into a sub-package
