+++ doc/pipelinemodules/Pipeline/Control.rst
@@ -1,0 +1,4 @@
+
+.. automodule:: Pipeline.Control
+   :members:
+   :show-inheritance:
