+++ doc/pipelinemodules/Pipeline/Database.rst
@@ -1,0 +1,4 @@
+
+.. automodule:: Pipeline.Database
+   :members:
+   :show-inheritance:
