+++ doc/pipelinemodules/Pipeline/Execution.rst
@@ -1,0 +1,4 @@
+
+.. automodule:: Pipeline.Execution
+   :members:
+   :show-inheritance:
