+++ doc/pipelinemodules/Pipeline/Local.rst
@@ -1,0 +1,4 @@
+
+.. automodule:: Pipeline.Local
+   :members:
+   :show-inheritance:
