+++ doc/pipelinemodules/Pipeline/Utils.rst
@@ -1,0 +1,4 @@
+
+.. automodule:: Pipeline.Utils
+   :members:
+   :show-inheritance:
